+++ basis/Holmakefile
@@ -2,13 +2,10 @@ INCLUDES = ../misc pure ../translator ../characteristic ../semantics
 OPTIONS = QUIT_ON_FAILURE
 THYFILES = $(patsubst %Script.sml,%Theory.uo,$(wildcard *.sml))
 TARGETS = $(patsubst %.sml,%.uo,$(THYFILES))
-all: $(TARGETS) basis_ffi.o fs_ffi.o
+all: $(TARGETS) basis_ffi.o
 .PHONY: all
 ifndef CC
 CC=gcc
 endif
 basis_ffi.o: basis_ffi.c
-	$(CC) -c $< -o $@
-
-fs_ffi.o: fs_ffi.c
 	$(CC) -c $< -o $@
