+++ basis/basis_ffi.c
@@ -1,38 +1,7 @@
 #include <stdio.h>
 #include <string.h>
-
-/* stdout */
-
-void ffiputChar (char* c, long clen, char* a) {
-  putchar(a[0]);
-}
-
-// size_t is guaranteed to be at least 16 bytes.
-#define WRITE_MAXLEN 65535
-
-void ffiwriteStr (char* c, long clen, char* a) {
-  fwrite(c,
-         sizeof(char),
-         clen <= WRITE_MAXLEN ? clen : WRITE_MAXLEN,
-         stdout);
-}
-
-/* stderr */
-void ffiputChar_err(char* c, long clen, char* a, long len) {
-  putc(a[0], stderr);
-}
-
-/* stdin */
-
-void ffigetChar (char* conf, long clen, char* a) {
-  int c = getchar();
-  if(c == EOF) {
-    a[1] = 1;
-  } else {
-    a[0] = c;
-    a[1] = 0;
-  }
-}
+#include <unistd.h>
+#include <fcntl.h>
 
 /* commandLine */
 
@@ -42,7 +11,7 @@ extern char **argv;
 
 #define MAXLEN 256
 
-void ffigetArgs (char* c, long clen, char *a) {
+void ffigetArgs (unsigned char *c, long clen, unsigned char *a, long alen) {
         int i, j, k;
 
         for (i = 0, k = 0; (i < argc) && (k < MAXLEN); i++, k++) {
@@ -54,50 +23,71 @@ void ffigetArgs (char* c, long clen, char *a) {
         return;
 }
 
-/* rofs (read-only file system) */
-
-FILE* infds[256];
+/* 0 indicates null fd */
+int infds[256] = {STDIN_FILENO, STDOUT_FILENO, STDERR_FILENO, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1};
 
 int nextFD() {
   int fd = 0;
-  while(fd < 256 && infds[fd] != NULL) fd++;
+  while(fd < 256 && infds[fd] != -1) fd++;
   return fd;
 }
 
-void ffiopen (char* c, long clen, char *a) {
+void ffiopen_in (unsigned char *c, long clen, unsigned char *a, long alen) {
   int fd = nextFD();
-  if (fd < 255 && (infds[fd] = fopen(a,"r")))
-    a[0] = fd;
+  if (fd <= 255 && (infds[fd] = open(a, O_RDONLY))){
+    a[0] = 0;
+    a[1] = fd;
+  }
   else
     a[0] = 255;
 }
 
-void ffifgetc (char* conf, long clen, char *a) {
-  int c; /* not char, otherwise EOF is mapped to a valid char */
-  if (infds[a[0]] && (c = fgetc(infds[a[0]])) != EOF)
-    a[0] = c;
+void ffiopen_out (unsigned char *c, long clen, unsigned char *a, long alen) {
+  int fd = nextFD();
+  if (fd <= 255 && (infds[fd] = open(a, O_RDWR|O_CREAT|O_TRUNC))){
+    a[0] = 0;
+    a[1] = fd;
+  }
   else
     a[0] = 255;
 }
 
-void fficlose (char* c, long clen, char *a) {
-  if (infds[a[0]] && fclose(infds[a[0]]) == 0) {
-    infds[a[0]] = NULL;
+void ffiread (unsigned char *c, long clen, unsigned char *a, long alen) {
+  int nread = read(infds[a[0]], &a[3], a[1]);
+  if(nread < 0){
+    a[0] = 1;
+  }
+  else{
+    a[0] = 0; 
+    a[1] = nread;
+  }
+}
+
+void ffiwrite (unsigned char *c, long clen, unsigned char *a, long alen){
+  int nw = write(infds[a[0]], &a[3+a[2]], a[1]);
+  if(nw < 0){
+      a[0] = 1;
+  }
+
+  else{
+    a[0] = 0; 
+    a[1] = nw;
+  }
+}
+
+void fficlose (unsigned char *c, long clen, unsigned char *a, long alen) {
+  if (infds[a[0]] && close(infds[a[0]]) == 0) {
+    infds[a[0]] = -1;
     a[0] = 1;
   }
   else
     a[0] = 0;
 }
-
-void ffiisEof (char* conf, long clen, char *a) {
-  int c; /* not char, otherwise EOF is mapped to a valid char */
-  if (infds[a[0]])
-    if ((c = fgetc(infds[a[0]])) == EOF)
-      a[0] = 1;
-    else {
-      ungetc(c, infds[a[0]]);
-      a[0] = 0;
-    }
+/*
+void ffiseek (unsigned char *a) {
+  int off = lseek(infds[a[0]], a[2], SEEK_SET);
+  if (off = -1)
+    a[0] = 1;
   else
-    a[0] = 255;
-}
+    a[0] = 0;
+}*/
